Move `Reflected` to `SwiftagramTests`
