+++ Sources/Swiftagram/Models/Specialized/Friendship.swift
@@ -8,19 +8,7 @@
 import Foundation
 
 /// A `struct` representing a `Friendship`.
-public struct Friendship: ReflectedType {
-    /// The debug description prefix.
-    public static let debugDescriptionPrefix: String = ""
-    /// A list of to-be-reflected properties.
-    public static let properties: [String: PartialKeyPath<Self>] = ["isFollowedByYou": \Self.isFollowedByYou,
-                                                                    "isFollowingYou": \Self.isFollowingYou,
-                                                                    "isBlockedByYou": \Self.isBlockedByYou,
-                                                                    "isCloseFriend": \Self.isCloseFriend,
-                                                                    "didRequestToFollowYou": \Self.didRequestToFollowYou,
-                                                                    "didRequestToFollow": \Self.didRequestToFollow,
-                                                                    "isMutingStories": \Self.isMutingStories,
-                                                                    "isMutingPosts": \Self.isMutingPosts]
-
+public struct Friendship: Wrapped {
     /// The underlying `Response`.
     public var wrapper: () -> Wrapper
 
@@ -51,13 +39,7 @@ public struct Friendship: ReflectedType {
 
 public extension Friendship {
     /// A `struct` representing a `Friendship` collection.
-    struct Dictionary: ResponseType, ReflectedType {
-        /// The prefix.
-        public static var debugDescriptionPrefix: String { "Friendship." }
-        /// A list of to-be-reflected properties.
-        public static let properties: [String: PartialKeyPath<Self>] = ["friendships": \Self.friendships,
-                                                                        "error": \Self.error]
-
+    struct Dictionary: ResponseType {
         /// The underlying `Response`.
         public var wrapper: () -> Wrapper
