+++ Sources/Swiftagram/Models/Specialized/Media.swift
@@ -9,17 +9,9 @@ import CoreGraphics
 import Foundation
 
 /// A `struct` representing a `Media`.
-public struct Media: ReflectedType {
+public struct Media: Wrapped {
     /// A `struct` representing some content `Version`.
-    public struct Version: ReflectedType {
-        /// The prefix.
-        public static var debugDescriptionPrefix: String { "Media." }
-        /// A list of to-be-reflected properties.
-        public static let properties: [String: PartialKeyPath<Self>] = ["url": \Self.url,
-                                                                        "size": \Self.size,
-                                                                        "aspectRatio": \Self.aspectRatio,
-                                                                        "resolution": \Self.resolution]
-
+    public struct Version: Wrapped {
         /// The underlying `Response`.
         public var wrapper: () -> Wrapper
 
@@ -45,12 +37,7 @@ public struct Media: ReflectedType {
     }
 
     /// A `struct` representing a `Picture`.
-    public struct Picture: ReflectedType {
-        /// The prefix.
-        public static var debugDescriptionPrefix: String { "Media." }
-        /// A list of to-be-reflected properties.
-        public static let properties: [String: PartialKeyPath<Self>] = ["images": \Self.images]
-
+    public struct Picture: Wrapped {
         /// The underlying `Response`.
         public var wrapper: () -> Wrapper
 
@@ -70,14 +57,7 @@ public struct Media: ReflectedType {
     }
 
     /// A `struct` representing a `Video`.
-    public struct Video: ReflectedType {
-        /// The prefix.
-        public static var debugDescriptionPrefix: String { "Media." }
-        /// A list of to-be-reflected properties.
-        public static let properties: [String: PartialKeyPath<Self>] = ["duration": \Self.duration,
-                                                                        "images": \Self.images,
-                                                                        "clips": \Self.clips]
-
+    public struct Video: Wrapped {
         /// The underlying `Response`.
         public var wrapper: () -> Wrapper
 
@@ -145,25 +125,6 @@ public struct Media: ReflectedType {
         }
     }
 
-    /// The debug description prefix.
-    public static let debugDescriptionPrefix: String = ""
-    /// A list of to-be-reflected properties.
-    public static let properties: [String: PartialKeyPath<Self>] = ["identifier": \Self.identifier,
-                                                                    "primaryKey": \Self.primaryKey,
-                                                                    "code": \Self.code,
-                                                                    "wasLikedByYou": \Self.wasLikedByYou,
-                                                                    "expiringAt": \Self.expiringAt,
-                                                                    "takenAt": \Self.takenAt,
-                                                                    "size": \Self.size,
-                                                                    "aspectRatio": \Self.aspectRatio,
-                                                                    "resolution": \Self.resolution,
-                                                                    "caption": \Self.caption,
-                                                                    "comments": \Self.comments,
-                                                                    "likes": \Self.likes,
-                                                                    "content": \Self.content,
-                                                                    "user": \Self.user,
-                                                                    "location": \Self.location]
-
     /// The underlying `Response`.
     public var wrapper: () -> Wrapper
 
@@ -218,12 +179,7 @@ public struct Media: ReflectedType {
 
 public extension Media {
     /// A `struct` representing a `Media` single response.
-    struct Unit: ResponseType, ReflectedType {
-        /// The prefix.
-        public static var debugDescriptionPrefix: String { "Media." }
-        /// A list of to-be-reflected properties.
-        public static let properties: [String: PartialKeyPath<Self>] = ["media": \Self.media,
-                                                                        "error": \Self.error]
+    struct Unit: ResponseType {
         /// The underlying `Response`.
         public var wrapper: () -> Wrapper
 
@@ -240,15 +196,9 @@ public extension Media {
     }
 
     /// A `struct` representing a `Media` collection.
-    struct Collection: ResponseType, ReflectedType, PaginatedType {
+    struct Collection: ResponseType, Paginatable {
         /// The associated offset type.
         public typealias Offset = String?
-
-        /// The prefix.
-        public static var debugDescriptionPrefix: String { "Media." }
-        /// A list of to-be-reflected properties.
-        public static let properties: [String: PartialKeyPath<Self>] = ["media": \Self.media,
-                                                                        "error": \Self.error]
 
         /// The underlying `Response`.
         public var wrapper: () -> Wrapper
