+++ Tests/SwiftagramTests/Reflection/Status.swift
@@ -1,0 +1,17 @@
+//
+//  Status.swift
+//  SwiftagramTests
+//
+//  Created by Stefano Bertagno on 20/03/21.
+//
+
+import Foundation
+
+import Swiftagram
+
+extension Status: Reflected {
+    /// The debug description prefix.
+    public static let debugDescriptionPrefix: String = ""
+    /// A list of to-be-reflected properties.
+    public static let properties: [String: PartialKeyPath<Self>] = ["error": \Self.error]
+}
