+++ Tests/SwiftagramTests/Reflection/Sticker.swift
@@ -1,0 +1,20 @@
+//
+//  Sticker.swift
+//  SwiftagramTests
+//
+//  Created by Stefano Bertagno on 20/03/21.
+//
+
+import Foundation
+
+import Swiftagram
+
+extension Sticker: Reflected {
+    /// The debug description prefix.
+    public static let debugDescriptionPrefix: String = ""
+    /// A list of to-be-reflected properties.
+    public static let properties: [String: PartialKeyPath<Self>] = ["identifier": \Self.identifier,
+                                                                    "level": \Self.level,
+                                                                    "offset": \Self.offset,
+                                                                    "rotation": \Self.rotation]
+}
